+++ silo-core/contracts/interfaces/ISiloFactory.sol
@@ -49,9 +49,20 @@ interface ISiloFactory is IERC721 {
     /// @param daoFeeReceiver Address of the new DAO fee receiver.
     event DaoFeeReceiverChanged(address daoFeeReceiver);
 
+    /// @notice Emitted on the change of DAO fee receiver for particular silo
+    /// @param silo Address for which new DAO fee receiver is set.
+    /// @param daoFeeReceiver Address of the new DAO fee receiver.
+    event DaoFeeReceiverChangedForSilo(address silo, address daoFeeReceiver);
+
+    /// @notice Emitted on the change of DAO fee receiver for particular asset
+    /// @param asset Address for which new DAO fee receiver is set.
+    /// @param daoFeeReceiver Address of the new DAO fee receiver.
+    event DaoFeeReceiverChangedForAsset(address asset, address daoFeeReceiver);
+
     error MissingHookReceiver();
     error ZeroAddress();
     error DaoFeeReceiverZeroAddress();
+    error SameDaoFeeReceiver();
     error EmptyToken0();
     error EmptyToken1();
     error MaxFeeExceeded();
@@ -73,19 +84,25 @@ interface ISiloFactory is IERC721 {
     error HookIsZeroAddress();
     error LiquidationTargetLtvTooHigh();
     error NotYourSilo();
+    error ConfigMismatchSilo();
+    error ConfigMismatchShareProtectedToken();
+    error ConfigMismatchShareDebtToken();
+    error ConfigMismatchShareCollateralToken();
 
     /// @notice Create a new Silo.
-    /// @param _initData Silo initialization data.
     /// @param _siloConfig Silo configuration.
     /// @param _siloImpl Address of the `Silo` implementation.
     /// @param _shareProtectedCollateralTokenImpl Address of the `ShareProtectedCollateralToken` implementation.
     /// @param _shareDebtTokenImpl Address of the `ShareDebtToken` implementation.
+    /// @param _deployer Address of the deployer.
+    /// @param _creator Address of the creator.
     function createSilo(
-        ISiloConfig.InitData memory _initData,
         ISiloConfig _siloConfig,
         address _siloImpl,
         address _shareProtectedCollateralTokenImpl,
-        address _shareDebtTokenImpl
+        address _shareDebtTokenImpl,
+        address _deployer,
+        address _creator
     )
         external;
 
@@ -100,9 +117,19 @@ interface ISiloFactory is IERC721 {
     /// @param _maxFee Value of the new DAO maximal fee.
     function setDaoFee(uint128 _minFee, uint128 _maxFee) external;
 
-    /// @notice Set the new DAO fee receiver.
+    /// @notice Set the default DAO fee receiver.
     /// @param _newDaoFeeReceiver Address of the new DAO fee receiver.
     function setDaoFeeReceiver(address _newDaoFeeReceiver) external;
+
+    /// @notice Set the new DAO fee receiver for asset, this setup will be used when fee receiver for silo is empty.
+    /// @param _asset Address for which new DAO fee receiver is set.
+    /// @param _newDaoFeeReceiver Address of the new DAO fee receiver.
+    function setDaoFeeReceiverForAsset(address _asset, address _newDaoFeeReceiver) external;
+
+    /// @notice Set the new DAO fee receiver for silo. This setup has highest priority.
+    /// @param _silo Address for which new DAO fee receiver is set.
+    /// @param _newDaoFeeReceiver Address of the new DAO fee receiver.
+    function setDaoFeeReceiverForSilo(address _silo, address _newDaoFeeReceiver) external;
 
     /// @notice Update the value of max deployer fee. Updated value will be used only for a new Silos max deployer
     /// fee validation. Previously deployed SiloConfigs are immutable.
@@ -140,6 +167,9 @@ interface ISiloFactory is IERC721 {
 
     /// @notice Get SiloConfig address by Silo id.
     function idToSiloConfig(uint256 _id) external view returns (address);
+
+    /// @notice Get the counter of silos created by the wallet.
+    function creatorSiloCounter(address _creator) external view returns (uint256);
 
     /// @notice Do not use this method to check if silo is secure. Anyone can deploy silo with any configuration
     /// and implementation. Most critical part of verification would be to check who deployed it.