+++ silo-vaults/contracts/SiloVault.sol
@@ -11,6 +11,8 @@ import {ERC20} from "openzeppelin5/token/ERC20/ERC20.sol";
 import {SafeERC20} from "openzeppelin5/token/ERC20/utils/SafeERC20.sol";
 import {UtilsLib} from "morpho-blue/libraries/UtilsLib.sol";
 
+import {TokenHelper} from "silo-core/contracts/lib/TokenHelper.sol";
+
 import {
     MarketConfig,
     PendingUint192,
@@ -24,11 +26,11 @@ import {INotificationReceiver} from "./interfaces/INotificationReceiver.sol";
 import {IVaultIncentivesModule} from "./interfaces/IVaultIncentivesModule.sol";
 import {IIncentivesClaimingLogic} from "./interfaces/IIncentivesClaimingLogic.sol";
 
-
 import {PendingUint192, PendingAddress, PendingLib} from "./libraries/PendingLib.sol";
 import {ConstantsLib} from "./libraries/ConstantsLib.sol";
 import {ErrorsLib} from "./libraries/ErrorsLib.sol";
 import {EventsLib} from "./libraries/EventsLib.sol";
+import {SiloVaultActionsLib} from "./libraries/SiloVaultActionsLib.sol";
 
 /// @title SiloVault
 /// @dev Forked with gratitude from Morpho Labs.
@@ -85,9 +87,6 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
     address public feeRecipient;
 
     /// @inheritdoc ISiloVaultBase
-    address public skimRecipient;
-
-    /// @inheritdoc ISiloVaultBase
     IERC4626[] public supplyQueue;
 
     /// @inheritdoc ISiloVaultBase
@@ -118,7 +117,9 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
         require(_asset != address(0), ErrorsLib.ZeroAddress());
         require(address(_vaultIncentivesModule) != address(0), ErrorsLib.ZeroAddress());
 
-        DECIMALS_OFFSET = uint8(UtilsLib.zeroFloorSub(18, IERC20Metadata(_asset).decimals()));
+        uint256 decimals = TokenHelper.assertAndGetDecimals(_asset);
+        require(decimals <= 18, ErrorsLib.NotSupportedDecimals());
+        DECIMALS_OFFSET = uint8(UtilsLib.zeroFloorSub(18 + 3, decimals));
 
         _checkTimelockBounds(_initialTimelock);
         _setTimelock(_initialTimelock);
@@ -185,20 +186,7 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
 
     /// @inheritdoc ISiloVaultBase
     function setIsAllocator(address _newAllocator, bool _newIsAllocator) external virtual onlyOwner {
-        if (isAllocator[_newAllocator] == _newIsAllocator) revert ErrorsLib.AlreadySet();
-
-        isAllocator[_newAllocator] = _newIsAllocator;
-
-        emit EventsLib.SetIsAllocator(_newAllocator, _newIsAllocator);
-    }
-
-    /// @inheritdoc ISiloVaultBase
-    function setSkimRecipient(address _newSkimRecipient) external virtual onlyOwner {
-        if (_newSkimRecipient == skimRecipient) revert ErrorsLib.AlreadySet();
-
-        skimRecipient = _newSkimRecipient;
-
-        emit EventsLib.SetSkimRecipient(_newSkimRecipient);
+        SiloVaultActionsLib.setIsAllocator(_newAllocator, _newIsAllocator, isAllocator);
     }
 
     /// @inheritdoc ISiloVaultBase
@@ -490,20 +478,10 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
     }
 
     /// @inheritdoc ISiloVaultBase
-    function skim(address _token) external virtual {
-        if (skimRecipient == address(0)) revert ErrorsLib.ZeroAddress();
-
-        uint256 amount = _ERC20BalanceOf(_token, address(this));
-
-        IERC20(_token).safeTransfer(skimRecipient, amount);
-
-        emit EventsLib.Skim(_msgSender(), _token, amount);
-    }
-
-    /// @inheritdoc ISiloVaultBase
     function claimRewards() public virtual {
         _nonReentrantOn();
 
+        _updateLastTotalAssets(_accrueFee());
         _claimRewards();
 
         _nonReentrantOff();
@@ -518,7 +496,7 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
 
     /// @inheritdoc IERC20Metadata
     function decimals() public view virtual override(ERC20, ERC4626) returns (uint8) {
-        return ERC4626.decimals();
+        return 18;
     }
 
     /// @inheritdoc IERC4626
@@ -551,6 +529,29 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
         return _convertToSharesWithTotals(assets, newTotalSupply, newTotalAssets, Math.Rounding.Floor);
     }
 
+    /// @inheritdoc IERC20
+    function transfer(address _to, uint256 _value) public virtual override(ERC20, IERC20) returns (bool success) {
+        _nonReentrantOn();
+
+        success = ERC20.transfer(_to, _value);
+
+        _nonReentrantOff();
+    }
+
+    /// @inheritdoc IERC20
+    function transferFrom(address _from, address _to, uint256 _value)
+        public
+        virtual
+        override(ERC20, IERC20)
+        returns (bool success)
+    {
+        _nonReentrantOn();
+
+        success = ERC20.transferFrom(_from, _to, _value);
+
+        _nonReentrantOff();
+    }
+
     /// @inheritdoc IERC4626
     function deposit(uint256 _assets, address _receiver) public virtual override returns (uint256 shares) {
         _nonReentrantOn();
@@ -561,9 +562,10 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
         // It is updated again in `_deposit`.
         lastTotalAssets = newTotalAssets;
 
-        shares = _convertToSharesWithTotals(_assets, totalSupply(), newTotalAssets, Math.Rounding.Floor);
+        shares = _convertToSharesWithTotalsSafe(_assets, totalSupply(), newTotalAssets, Math.Rounding.Floor);
 
         _deposit(_msgSender(), _receiver, _assets, shares);
+        _assetLossCheck(shares, _assets);
 
         _nonReentrantOff();
     }
@@ -578,9 +580,10 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
         // It is updated again in `_deposit`.
         lastTotalAssets = newTotalAssets;
 
-        assets = _convertToAssetsWithTotals(_shares, totalSupply(), newTotalAssets, Math.Rounding.Ceil);
+        assets = _convertToAssetsWithTotalsSafe(_shares, totalSupply(), newTotalAssets, Math.Rounding.Ceil);
 
         _deposit(_msgSender(), _receiver, assets, _shares);
+        _assetLossCheck(_shares, assets);
 
         _nonReentrantOff();
     }
@@ -598,7 +601,7 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
 
         // Do not call expensive `maxWithdraw` and optimistically withdraw assets.
 
-        shares = _convertToSharesWithTotals(_assets, totalSupply(), newTotalAssets, Math.Rounding.Ceil);
+        shares = _convertToSharesWithTotalsSafe(_assets, totalSupply(), newTotalAssets, Math.Rounding.Ceil);
 
         // `newTotalAssets - assets` may be a little off from `totalAssets()`.
         _updateLastTotalAssets(UtilsLib.zeroFloorSub(newTotalAssets, _assets));
@@ -620,7 +623,7 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
 
         // Do not call expensive `maxRedeem` and optimistically redeem shares.
 
-        assets = _convertToAssetsWithTotals(_shares, totalSupply(), newTotalAssets, Math.Rounding.Floor);
+        assets = _convertToAssetsWithTotalsSafe(_shares, totalSupply(), newTotalAssets, Math.Rounding.Floor);
 
         // `newTotalAssets - assets` may be a little off from `totalAssets()`.
         _updateLastTotalAssets(UtilsLib.zeroFloorSub(newTotalAssets, assets));
@@ -658,7 +661,7 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
         newTotalSupply = totalSupply() + feeShares;
 
         assets = _convertToAssetsWithTotals(balanceOf(_owner), newTotalSupply, newTotalAssets, Math.Rounding.Floor);
-        assets -= _simulateWithdrawERC4626(assets);
+        assets -= SiloVaultActionsLib.simulateWithdrawERC4626(assets, withdrawQueue);
     }
 
     /// @dev Returns the maximum amount of assets that the vault can supply to ERC4626 vaults.
@@ -703,6 +706,19 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
         return _assets.mulDiv(_newTotalSupply + 10 ** _decimalsOffset(), _newTotalAssets + 1, _rounding);
     }
 
+    /// @dev Returns the amount of shares that the vault would exchange for the amount of `assets` provided.
+    /// @dev It assumes that the arguments `newTotalSupply` and `newTotalAssets` are up to date.
+    /// @dev Reverts if the result is zero.
+    function _convertToSharesWithTotalsSafe(
+        uint256 _assets,
+        uint256 _newTotalSupply,
+        uint256 _newTotalAssets,
+        Math.Rounding _rounding
+    ) internal view virtual returns (uint256 shares) {
+        shares = _convertToSharesWithTotals(_assets, _newTotalSupply, _newTotalAssets, _rounding);
+        require(shares != 0, ErrorsLib.ZeroShares());
+    }
+
     /// @dev Returns the amount of assets that the vault would exchange for the amount of `shares` provided.
     /// @dev It assumes that the arguments `newTotalSupply` and `newTotalAssets` are up to date.
     function _convertToAssetsWithTotals(
@@ -712,6 +728,19 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
         Math.Rounding _rounding
     ) internal view virtual returns (uint256) {
         return _shares.mulDiv(_newTotalAssets + 1, _newTotalSupply + 10 ** _decimalsOffset(), _rounding);
+    }
+
+    /// @dev Returns the amount of assets that the vault would exchange for the amount of `shares` provided.
+    /// @dev It assumes that the arguments `newTotalSupply` and `newTotalAssets` are up to date.
+    /// @dev Reverts if the result is zero.
+    function _convertToAssetsWithTotalsSafe(
+        uint256 _shares,
+        uint256 _newTotalSupply,
+        uint256 _newTotalAssets,
+        Math.Rounding _rounding
+    ) internal view virtual returns (uint256 assets) {
+        assets = _convertToAssetsWithTotals(_shares, _newTotalSupply, _newTotalAssets, _rounding);
+        require(assets != 0, ErrorsLib.ZeroAssets());
     }
 
     /// @inheritdoc ERC4626
@@ -756,7 +785,7 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
     {
         shares = _ERC20BalanceOf(address(_market), address(this));
         // we assume here, that in case of any interest on IERC4626, convertToAssets returns assets with interest
-        assets = _market.convertToAssets(shares);
+        assets = _previewRedeem(_market, shares);
     }
 
     /// @dev Reverts if `newTimelock` is not within the bounds.
@@ -786,6 +815,7 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
     /// @dev Sets the cap of the market.
     function _setCap(IERC4626 _market, uint184 _supplyCap) internal virtual {
         MarketConfig storage marketConfig = config[_market];
+        uint256 approveValue;
 
         if (_supplyCap > 0) {
             if (!marketConfig.enabled) {
@@ -802,11 +832,13 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
             }
 
             marketConfig.removableAt = 0;
+            // one time approval, so market can pull any amount of tokens from SiloVault in a future
+            approveValue = type(uint256).max;
         }
 
         marketConfig.cap = _supplyCap;
-        // one time approval, so market can pull any amount of tokens from SiloVault in a future
-        IERC20(asset()).forceApprove(address(_market), type(uint256).max);
+        IERC20(asset()).forceApprove(address(_market), approveValue);
+
         emit EventsLib.SetCap(_msgSender(), _market, _supplyCap);
 
         delete pendingCap[_market];
@@ -864,20 +896,6 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
         if (_assets != 0) revert ErrorsLib.NotEnoughLiquidity();
     }
 
-    /// @dev Simulates a withdraw of `assets` from ERC4626 vault.
-    /// @return The remaining assets to be withdrawn.
-    function _simulateWithdrawERC4626(uint256 _assets) internal view virtual returns (uint256) {
-        for (uint256 i; i < withdrawQueue.length; ++i) {
-            IERC4626 market = withdrawQueue[i];
-
-            _assets = UtilsLib.zeroFloorSub(_assets, market.maxWithdraw(address(this)));
-
-            if (_assets == 0) break;
-        }
-
-        return _assets;
-    }
-
     /* FEE MANAGEMENT */
 
     /// @dev Updates `lastTotalAssets` to `updatedTotalAssets`.
@@ -916,7 +934,7 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
 
     /// @notice Returns the expected supply assets balance of `user` on a market after having accrued interest.
     function _expectedSupplyAssets(IERC4626 _market, address _user) internal view virtual returns (uint256 assets) {
-        assets = _market.convertToAssets(_ERC20BalanceOf(address(_market), _user));
+        assets = _previewRedeem(_market, _ERC20BalanceOf(address(_market), _user));
     }
 
     function _update(address _from, address _to, uint256 _value) internal virtual override {
@@ -978,5 +996,21 @@ contract SiloVault is ERC4626, ERC20Permit, Ownable2Step, Multicall, ISiloVaultS
     /// @dev to save code size ~500 B
     function _ERC20BalanceOf(address _token, address _account) internal view returns (uint256 balance) {
         balance = IERC20(_token).balanceOf(_account);
+    }
+
+    function _previewRedeem(IERC4626 _market, uint256 _shares) internal view returns (uint256 assets) {
+        assets = _market.previewRedeem(_shares);
+    }
+
+    function _assetLossCheck(uint256 _shares, uint256 _expectedAssets) internal {
+        uint256 preview = previewRedeem(_shares);
+        if (preview >= _expectedAssets) return;
+
+        uint256 loss;
+        // save because we checking above `if (preview >= _expectedAssets)`
+        unchecked { loss = _expectedAssets - preview; }
+        uint256 arbitraryLossThreshold = 10;
+
+        require(loss < arbitraryLossThreshold, ErrorsLib.AssetLoss(loss));
     }
 }