+++ silo-vaults/test/foundry/incentives/VaultRewardsIntegrationSetup.sol
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-2.0-or-later
 pragma solidity ^0.8.28;
 
+import {IERC4626} from "openzeppelin5/interfaces/IERC4626.sol";
 import {Ownable} from "openzeppelin5/access/Ownable2Step.sol";
 import {Strings} from "openzeppelin5/utils/Strings.sol";
 import {Hook} from "silo-core/contracts/lib/Hook.sol";
@@ -19,6 +20,7 @@ import {SiloIncentivesControllerCL} from "../../../contracts/incentives/claiming
 
 import {INotificationReceiver} from "../../../contracts/interfaces/INotificationReceiver.sol";
 import {IntegrationTest} from "../helpers/IntegrationTest.sol";
+import {IVaultIncentivesModule} from "silo-vaults/contracts/interfaces/IVaultIncentivesModule.sol";
 
 import {CAP} from "../helpers/BaseTest.sol";
 
@@ -30,9 +32,13 @@ contract VaultRewardsIntegrationSetup is IntegrationTest {
 
     SiloIncentivesControllerGaugeLike siloIncentivesController;
     SiloIncentivesController vaultIncentivesController;
+    IVaultIncentivesModule vaultIncentivesModule;
 
     function setUp() public virtual override {
         super.setUp();
+
+        vaultIncentivesModule = vault.INCENTIVES_MODULE();
+        assertTrue(address(vaultIncentivesModule) != address(0), "empty vaultIncentivesModule");
 
         _setCap(allMarkets[0], _cap());
         _setCap(allMarkets[1], _cap());
@@ -74,6 +80,11 @@ contract VaultRewardsIntegrationSetup is IntegrationTest {
         );
 
         vm.prank(OWNER);
-        vaultIncentivesModule.addIncentivesClaimingLogic(address(silo1), cl);
+        vaultIncentivesModule.submitIncentivesClaimingLogic(IERC4626(address(silo1)), cl);
+
+        vm.warp(block.timestamp + vault.timelock() + 1);
+
+        vm.prank(OWNER);
+        vaultIncentivesModule.acceptIncentivesClaimingLogic(IERC4626(address(silo1)), cl);
     }
 }